refactor(tui): decompose app/mod.rs into focused submodules (KAN-763)
